object: fix Go 1.25 lint for conn max age
